Used size_t instead of unsigned int

+++ clang/include/joe_Array.h
@@ -23,8 +23,8 @@
 typedef joe_Object joe_Array;
 extern joe_Class joe_Array_Class;
 
-JOEOBJ_API joe_Object joe_Array_New (unsigned int length);
-JOEOBJ_API unsigned int joe_Array_length (joe_Array self);
+JOEOBJ_API joe_Object joe_Array_New (size_t length);
+JOEOBJ_API size_t joe_Array_length (joe_Array self);
 JOEOBJ_API joe_Array joe_Array_clean(joe_Array self);
 JOEOBJ_API int joe_Array_add(joe_Object self, joe_Object obj, joe_Object *retval);
